+++ src/js/core/policies.js
@@ -1972,6 +1972,20 @@ const policies = {
     ]
   },
 
+  Preference_browser_cache_disk_enable : {
+    description : browser.i18n.getMessage('policy_description_Preference_browser_cache_disk_enable'),
+    enterprise_only : false,
+    first_available : { mainstream : '68.0', esr : '68.0' },
+    info_link : null,
+    ui_category : 'others',
+    type : 'preference',
+    properties : {
+      option : 'browser.cache.disk.enable',
+      type : 'boolean',
+      default : 'true'
+    }
+  },
+
   HardwareAcceleration : {
     description : browser.i18n.getMessage('policy_description_HardwareAcceleration'),
     enterprise_only : false,